Added alot of stuff

+++ Assets/Scripts/Items/Items/BundleOfDynamite.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Items/BundleOfDynamite")]
 public class BundleOfDynamite : Item
 {
-    [SerializeField] private int baseExplodeDamage;
+    [SerializeField] private float baseExplodeDamage;
     [SerializeField] private float baseBurnDamage;
     [SerializeField] private float baseExplodeRadius;
 
-    [SerializeField] private int stackExplodeDamage;
+    [SerializeField] private float stackExplodeDamage;
     [SerializeField] private float stackBurnDamage;
     [SerializeField] private float stackExplodeRadius;
